check on .update

+++ usr/bin/archboot-update.sh
@@ -107,5 +107,5 @@ if [[ -n "${_FULL_SYSTEM}" ]]; then
     _full_system | _dialog --title "${_MENU_TITLE}" --gauge "Refreshing pacman package database..." 6 75 0
     clear
 fi
-rm /.update
+[[ -e /.update ]] && rm /.update
 # vim: set ft=sh ts=4 sw=4 et:
